Added divisibility for number '7'
The code is now more flexible. All numbers for divisibility test are stored as an array. Therefore, any new number can be accomodated.
Also, the strings to be print for each number are in a list. Therefore, it is flexible in a sense it can add new/ change existing/ exchange strings without changing actual logic of calculating